+++ database/nostr-lmdb/CHANGELOG.md
@@ -33,7 +33,7 @@
 
 - Rename `NostrLMDB` to `NostrLmdb`
 - Spawn blocking task for read operations (https://github.com/rust-nostr/nostr/pull/1141)
-- Improve scraping performance (https://github.com/rust-nostr/nostr/pull/1221)
+- Improve scraping and counting performance (https://github.com/rust-nostr/nostr/pull/1221)
 
 ### Added
 
@@ -42,7 +42,7 @@
 
 ### Fixed
 
-- Fix the performance of queries by kinds (https://github.com/rust-nostr/nostr/pull/1220)
+- Fix the performance issues of queries by kinds (https://github.com/rust-nostr/nostr/pull/1220)
 
 ## v0.44.0 - 2025/11/06